requirements.txt
pandas
numpy
matplotlib
Pillow
python-dotenv
google-generativeai
elevenlabs
streamlit
mutagen
azure-cognitiveservices-vision-computervision
